refactor(ui): extract shared HeaderActions and navigation config
- Extract duplicated header actions JSX (~35 lines) into HeaderActions
  server component, used by both (pages)/layout.tsx and page.tsx
- Extract navigation array into shared navigation.ts, imported by
  both header.tsx and home-hero.tsx
- Remove unused Route type import from home-hero.tsx
- Simplify (pages)/layout.tsx from 75 to 17 lines
- Simplify page.tsx from 102 to 42 lines
