+++ src/app/(public)/_components/header.tsx
@@ -1,21 +1,11 @@
 import { MenuIcon } from "lucide-react";
-import type { Route } from "next";
 import Link from "next/link";
 import { type ReactNode, Suspense } from "react";
 import { Icons } from "@/components/icons";
 import { MobileNavigation } from "@/components/mobile-nav";
 import { Container } from "@/components/shared/container";
 import { Button, buttonVariants } from "@/components/ui/button";
-import { featureFlags } from "@/config/features";
-
-const navigation: { name: string; href: Route }[] = [
-  { name: "E-shop", href: "/e-shop" },
-  { name: "B2B", href: "/b2b" },
-  { name: "Predajne", href: "/predajne" },
-  ...(featureFlags.blog ? [{ name: "Blog", href: "/blog" as Route }] : []),
-];
-
-export const headerNavigation = navigation;
+import { navigation } from "./navigation";
 
 export function Header({ children }: { children: ReactNode }) {
   return (
